requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
scipy>=1.9
matplotlib>=3.7
pyts>=0.13
scikit-image>=0.20
PyYAML>=6.0
tqdm>=4.66
tensorflow>=2.15
